# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EntreLibros API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
...
    "/api/messages": {
      "get": {
        "summary": "List the authenticated user's conversations",
        "tags": ["Messaging"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/messages/contacts": {
      "get": {
        "summary": "Search visible contacts for a new conversation",
        "tags": ["Messaging"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 80
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Visible contacts ordered with followed users first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["contacts"],
                  "properties": {
                    "contacts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "name", "alias", "isFollowing"],
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "alias": {
                            "type": "string"
                          },
                          "isFollowing": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/messages/conversations": {
      "post": {
        "summary": "Create a private conversation",
        "tags": ["Messaging"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["participantId"],
                "properties": {
                  "participantId": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation created"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "description": "Localized validation error"
          }
        }
      }
    },
    "/api/messages/{conversationId}/books": {
      "get": {
        "summary": "Listar libros elegibles de una conversación",
        "tags": ["Messaging"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Libros propios y libros públicos de la contraparte"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "La persona no participa en la conversación"
          }
        }
      }
    },
    "/api/messages/{conversationId}/messages": {
      "get": {
        "summary": "Read authorized conversation history",
        "tags": ["Messaging"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered message page"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Localized authorization error"
          },
          "422": {
            "description": "Localized pagination error"
          }
        }
      },
      "post": {
        "summary": "Persist an idempotent conversation message",
        "tags": ["Messaging"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clientKey", "body"],
                "properties": {
                  "clientKey": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  },
                  "attachmentMetadata": {
                    "description": "Adjunto persistido del chat. Los libros y propuestas se validan contra la conversación y las publicaciones de sus participantes.",
                    "oneOf": [
                      {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "contentType": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer",
                            "minimum": 1
                          },
                          "name": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string",
                            "enum": ["book"]
                          },
                          "bookId": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "author": {
                            "type": "string"
                          },
                          "coverUrl": {
                            "type": "string"
                          },
                          "ownerId": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "key",
                          "contentType",
                          "size",
                          "kind",
                          "bookId",
                          "title",
                          "author",
                          "coverUrl"
                        ]
                      },
                      {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "contentType": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer",
                            "minimum": 1
                          },
                          "name": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string",
                            "enum": ["swap"]
                          },
                          "offered": {
                            "$ref": "#/components/schemas/MessageBookReference"
                          },
                          "requested": {
                            "$ref": "#/components/schemas/MessageBookReference"
                          },
                          "note": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "key",
                          "contentType",
                          "size",
                          "kind",
                          "offered",
                          "requested"
                        ]
                      },
                      {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "contentType": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer",
                            "minimum": 1
                          },
                          "name": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string",
                            "enum": ["agreement"]
                          },
                          "agreementId": {
                            "type": "integer"
                          },
                          "version": {
                            "type": "integer"
                          },
                          "event": {
                            "type": "string",
                            "enum": [
                              "proposal",
                              "counterproposal",
                              "confirm",
                              "cancel",
                              "reject",
                              "complete"
                            ]
                          },
                          "details": {
                            "type": "object",
                            "properties": {
                              "meetingPoint": {
                                "type": "string"
                              },
                              "area": {
                                "type": "string"
                              },
                              "date": {
                                "type": "string"
                              },
                              "time": {
                                "type": "string"
                              },
                              "bookTitle": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "meetingPoint",
                              "area",
                              "date",
                              "time",
                              "bookTitle"
                            ]
                          },
                          "listingIds": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "actorName": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "key",
                          "contentType",
                          "size",
                          "kind",
                          "agreementId",
                          "version",
                          "event",
                          "details",
                          "listingIds",
                          "actorName"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Persisted message"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Localized authorization error"
          },
          "422": {
            "description": "Localized validation error"
          }
        }
      }
    },
    "/api/messages/{conversationId}/read": {
      "patch": {
        "summary": "Advance the authenticated participant read cursor",
        "tags": ["Messaging"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Read cursor updated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Localized authorization error"
          },
          "422": {
            "description": "Localized sequence error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionToken"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication is required"
      }
    },
    "schemas": {
      "MessageBookReference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          },
          "ownerId": {
            "type": "integer"
          }
        },
        "required": ["id", "title", "author", "coverUrl"]
      },
...
    }
  },
...
}